backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
sqlalchemy[asyncio]>=2.0.28
aiosqlite>=0.20.0
asyncpg>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
openai>=1.14.0
azure-ai-formrecognizer>=3.3.3
azure-core>=1.24.0,<1.30.0
tavily-python>=0.3.3
beautifulsoup4>=4.12.3
python-docx>=1.1.0
reportlab>=4.1.0
Pillow>=10.2.0
python-dotenv>=1.0.1
pydantic>=2.6.3
pydantic-settings>=2.2.1
alembic>=1.13.1
numpy>=1.26.4
protobuf>=4.25.3,<5.0.0
